requirements.txt
numpy
torch
Pillow
boto3==1.34.148
requests
botocore
